requirements.txt
pillow
python-dotenv
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
google-generativeai
tavily-python
gtts
langchain
langchain-core
langchain-google-genai